+++ wcm/fileopers.cpp
@@ -865,9 +865,9 @@ NCNumberWin::~NCNumberWin() {}
 
 class NCProgressWin: public Win
 {
-	int64 _from, _to;
-	int64 _num;
-	int _lastWidth, _lastPos;
+	int64_t _from, _to;
+	int64_t _num;
+	int64_t _lastWidth, _lastPos;
 public:
 	NCProgressWin( Win* parent )
 		:  Win( Win::WT_CHILD, 0, parent, 0 ), _from( 0 ), _to( 0 ), _num( 0 ), _lastWidth( 0 ), _lastPos( 0 )
@@ -893,8 +893,8 @@ public:
 
 		if ( _num < _from || _to <= _from ) { return; }
 
-		int64 size = _to - _from;
-		int n = ( _lastWidth * _num ) / size;
+		int64_t size = _to - _from;
+		int64_t n = ( _lastWidth * _num ) / size;
 
 		if ( diapazonChanged || _lastPos != n )
 		{
@@ -946,8 +946,8 @@ void NCProgressWin::Paint( wal::GC& gc, const crect& paintRect )
 
 	if ( !( _num < _from || _to <= _from || w <= 0 ) )
 	{
-		int64 size = _to - _from;
-		int n = ( w * _num ) / size;
+		int64_t size = _to - _from;
+		int n = int( ( w * _num ) / size );
 
 		crect r = rect;
 		r.right = n;
